// web/src/kiosk/components/__generated__/ScanControllerSignOutMutation.graphql.ts
export type ScanControllerSignOutMutation$variables = {
  categoryId: string;
  endTime: number;
  id: string;
  startTime: number;
};
export type ScanControllerSignOutMutation$data = {
  readonly scanSignOut: {
    readonly category: {
      readonly id: string;
      readonly name: string;
    } | null | undefined;
    readonly endTime: number | null | undefined;
    readonly id: string;
    readonly person: {
      readonly firstName: string;
      readonly id: string;
      readonly lastName: string;
    };
    readonly startTime: number;
  };
};
export type ScanControllerSignOutMutation = {
  response: ScanControllerSignOutMutation$data;
  variables: ScanControllerSignOutMutation$variables;
};